requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
joblib
shap
matplotlib
pytest
xgboost
mlflow
pyyaml
